requirements.txt
numpy>=1.16.4
scipy
Pillow
cython
matplotlib
scikit-image
tensorflow>=1.3.0
keras>=2.0.8
opencv-python
h5py
imgaug
IPython[all]
scikit-image>=0.16.2
opencv-python>=4.6.0.66
imgaug>=0.2.9